chore(meshconfig): move some types
