+++ plugins/lime-plugin-mesh-wide-config/src/meshConfigPage.tsx
@@ -8,7 +8,7 @@ import {
     ConfigSection,
 } from "plugins/lime-plugin-mesh-wide-config/src/components/ConfigSection";
 import { MeshStatus } from "plugins/lime-plugin-mesh-wide-config/src/components/MeshStatus";
-import { useMeshWideConfig } from "plugins/lime-plugin-mesh-wide-config/src/meshWideQueries";
+import { useMeshWideConfig } from "plugins/lime-plugin-mesh-wide-config/src/meshConfigQueries";
 
 const MeshConfigPage = () => {
     const { data: meshWideConfig, isLoading } = useMeshWideConfig({});
