+++ plugins/lime-plugin-mesh-wide-config/src/meshConfigTypes.tsx
@@ -1,0 +1,6 @@
+export interface IMeshWideSection {
+    name: string;
+    options: { [key: string]: string };
+}
+
+export type IMeshWideConfig = IMeshWideSection[];
