+++ plugins/lime-plugin-mesh-wide/src/meshWideMocks.tsx
@@ -2,7 +2,6 @@ import {
     IBaseLink,
     IBatmanLinks,
     ILinks,
-    IMeshWideConfig,
     INodes,
     IWifiLinks,
     LinkType,
@@ -438,33 +437,3 @@ export const nodes = (): INodes => {
 
     return { ...newState, ...newNode };
 };
-
-export const getMeshWideConfig = async () => meshWideConfig;
-//
-const options = {
-    primary_interface: "eth0",
-    main_ipv4_address: "10.170.128.0/16/17",
-};
-
-const meshWideConfig: IMeshWideConfig = [
-    {
-        name: "lime system",
-        options,
-    },
-    {
-        name: "lime network",
-        options,
-    },
-    {
-        name: "lime wifi",
-        options,
-    },
-    {
-        name: "generic_uci_config prometheus",
-        options,
-    },
-    {
-        name: "run_asset prometheus_enable",
-        options,
-    },
-];
